PCAN-RP1210 unplugged device fix
PEAKRP32 gets borked when trying to connect to an adapter that isn't plugged in, and can cause programs to stall or crash. This is the fix for that. Now it detects if the device isn't connected and returns INVALID_CLIENT_ID if it's borked.

Also, added getAPIName() to RP1210API class.